CRUD endereco completo

+++ src/main/java/com/farmacia/pharma_manager/backend/endereco/EnderecoRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
+    // Aqui podemos adicionar consultas personalizadas se necessário
 }
